Refactor: Extract UI into separate files
