+++ zeapp/desktop/src/main/kotlin/de/berlindroid/zekompanion/desktop/State.kt
@@ -1,0 +1,16 @@
+package de.berlindroid.zekompanion.desktop
+
+import androidx.compose.ui.graphics.ImageBitmap
+
+sealed class State {
+    data object Undecided : State()
+    data class EditNameBadge(
+        val name: String,
+        val contact: String,
+        val nameFontSize: Int,
+    ) : State()
+
+    data class EditImage(
+        val image: ImageBitmap,
+    ) : State()
+}
